Adding requirements.txt, changing variable types and general debug

+++ src/height_map_driver/scripts/height_map.py
@@ -3,7 +3,7 @@ import numpy as np
 
 class HeightMap:
 
-    def __init__(self, depth_map: np.array, boom_height: int):
+    def __init__(self, depth_map: np.array, boom_height: float):
         self.depth_map = depth_map
         self.boom_height = boom_height
